requirements.txt
azure-identity ~= 1.25, >= 1.25.0
click ~= 8.3, >= 8.3.0
msal[broker] ~= 1.34, >= 1.34.0
msgraph-sdk ~= 1.45, >= 1.45.0
requests ~= 2.32, >= 2.32.5
ruff ~= 0.13, >= 0.13.2
tzdata >= 2025.2
